The post [TensorFlow Performance: Loading Models](https://towardsdatascience.com/tensorflow-performance-loading-models-fb2d0dc340a3) revealed just how much the model format affects its load time. This graph is the punchline:

![Chart of load times for SavedModel vs HDF5 showing a drop from ~10s to ~2s](https://github.com/dchaley/deepcell-imaging/assets/352005/757a7c47-d066-4535-9af2-ead368906ccd)

We loaded the SavedModel format and simply resaved it as `.keras`. To reload it, we needed to reference the DeepCell layer type `Location2D` in the model loader.

The results are astonishing! From local benchmarking:

```
Before:

Loaded model in 12.3 s
Ran prediction in 1.25 s; success: True
After:

Loaded model in 0.84 s
Ran prediction in 1.19 s; success: True
```

A reduction of ~11.5s, or ~90%. 🎉 


This PR also cleans up some model fetching we were doing before the mesmer-in-pieces app became a proper module. Oops 🙈   Sorry @bnovotny 


WIP for: #262 
Open work: test & get load times on cloud