@dchaley 2024-07-11
Note: the [TF Keras docs](https://www.tensorflow.org/tutorials/keras/save_and_load#saving_custom_objects) recommend adding an [annotation](https://www.tensorflow.org/api_docs/python/tf/keras/utils/register_keras_serializable) to the type rather than specifying it at load.

We tried this but ... no dice 🎲   likely, we need to mess with the module internals somehow, or, we need it in the upstream DeepCell source.

I'm documenting this for posterity, however I'm not sure it's worth it to investigate further. If/when we use models with different custom objects, we may want to revisit. Basically, the tradeoff is whether or not it's ok to hardcode a list of supported custom objects. My bet: yes, for quite some time 🤔    (even if we retrain the model it'll be on new **data**, not a new model structure)