@dchaley 2024-07-11
LOL! It was HDF5 all along: not `.keras` (which was released in a later TensorFlow version)